gl: Refactor capabilities and add GLSL version detection support.
